Remove not needed tests
